Feat/running

+++ RunWithMe/presentation/src/main/java/com/side/runwithme/view/running_list/bottomsheet/practice/PracticeSettingClickListener.kt
@@ -3,5 +3,5 @@ package com.side.runwithme.view.running_list.bottomsheet.practice
 import com.side.runwithme.util.GOAL_TYPE
 
 interface PracticeSettingClickListener {
-    fun onItemClick(type: GOAL_TYPE ,amount: Int)
+    fun onItemClick(type: GOAL_TYPE ,amount: Long)
 }
